Update allocation of hot dipped galvanised UK production to sectors
The amount of UK-produced steel going to car bodies-in-white manufacturing was
reviewed to be too small. We had already assumed all imported hot-dipped
galvanised steel was going to this sector (15a), but the allocation of UK
production was based on the ISSB tables, which sent only a small amount to
sector 15a. We have revered to Jon Cullen's global estimates of where hot-dipped
galvanised steel is used.